+++ src/main/telemetry/crsf.c
@@ -792,7 +792,6 @@ static const crsfSensorDefinition_t crsfCustomSensorDefinitions[] =
     TLM_SENSOR(GPS_GROUNDSPEED,         0x1128,   200,  3000,    0,     U16),
     TLM_SENSOR(GPS_HOME_DISTANCE,       0x1129,   200,  3000,    0,     U16),
     TLM_SENSOR(GPS_HOME_DIRECTION,      0x112A,   200,  3000,    0,     S16),
-    TLM_SENSOR(GPS_FIX_TYPE,            0x112C,   500,  3000,    0,     U8),
 
     TLM_SENSOR(CPU_LOAD,                0x1141,   500,  3000,    10,    U8),
     TLM_SENSOR(SYS_LOAD,                0x1142,   500,  3000,    10,    U8),
@@ -800,16 +799,12 @@ static const crsfSensorDefinition_t crsfCustomSensorDefinitions[] =
 
     TLM_SENSOR(MODEL_ID,                0x1200,   200,  3000,    0,     U8),
     TLM_SENSOR(FLIGHT_MODE,             0x1201,   200,  3000,    0,     U16),
-    TLM_SENSOR(ARMING_FLAGS,            0x1202,   200,  3000,    0,     U8),
     TLM_SENSOR(ARMING_DISABLE_FLAGS,    0x1203,   200,  3000,    0,     U32),
 
-    TLM_SENSOR(PID_PROFILE,             0x1211,   200,  3000,    0,     U8),
-    TLM_SENSOR(RATES_PROFILE,           0x1212,   200,  3000,    0,     U8),
-    TLM_SENSOR(LED_PROFILE,             0x1213,   200,  3000,    0,     U8),
-    TLM_SENSOR(BATTERY_PROFILE,         0x1214,   200,  3000,    0,     U8),
-    TLM_SENSOR(TV_PROFILE,              0x1215,   200,  3000,    0,     U8),
-
     TLM_SENSOR(ADJFUNC,                 0x1220,   200,  3000,    0,     AdjFunc),
+
+    TLM_SENSOR(SYSTEM_STATUS,           0x1230,   100,  3000,    0,     U32),
+    TLM_SENSOR(SYSTEM_CONFIG,           0x1231,   500,  3000,    0,     U32),
 
     TLM_SENSOR(DEBUG_0,                 0xDB00,   100,  3000,    0,     S32),
     TLM_SENSOR(DEBUG_1,                 0xDB01,   100,  3000,    0,     S32),